Bugfixx in error handling routines and counting fixes.
Previous error handling since the last commit did not increment
the ERROR variable and therefore got away unnoticed.

Furthermore, even caught errors were not counted correctly as the
variable names were incorrect. FAILED was superfluous and used
instead of the correct ERRORS.

Lastly, when a file already exists it is counted as skipped and no
longer as an error.